+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changes
 
+* 0.7.9 in progress
+  * Fixes #28 by adding `-?` / `--query` option to explain what `clj-new` will attempt to do.
+  * Fixes #27 by adding `-e` / `--env` option to add "environment variables" that will be available to templates via the new `project-data` function; also standardizes the data passed to the `app`, `lib`, and `template` built-in templates.
 * 0.7.8 -- 2019-08-24
   * Fixes `-v` / `--verbose` option handling (again!).
   * Updates `org.clojure/test.check` to `"0.10.0"` and `tools.deps.alpha` to 0.7.541 (and add `slf4j-nop` as a dependency now that t.d.a has removed it).